Convert simulated microbiome data to TreeSummarizedExperiment(TreeSE) data container

This function converts simulated microbial community data to TreeSummarizedExperiment(TreeSummarizedExperiment) format.

convertToTreeSE(assay, output, ...)

Arguments

assay

is a matrix-like or list of matrix-like object. Rows refer to taxa and columns refer to samples.

output

character value for storing the matrix in TreeSummarizedExperiment (output = TreeSE).

...

: additional parameters to pass

Value

TreeSummarizedExperiment data object containing abundance matrix

Details

The abundance matrix from the simulation functions in miaSim can be converted to TreeSummarizedExperiment class object.

Storing the data in TreeSummarizedExperiment enables access to various Bioconductor packages and tools that extend the this class. This offers data and metadata synchronization, while accommodating specialized data structures for particular scientific applications.

Examples for TreeSE object manipulation and analysis can be found at https://microbiome.github.io/OMA

Examples

# Simulate time series data
x <- simulateHubbellRates(
    community_initial = c(0,5,10), migration_p = 0.01,
    metacommunity_p = NULL, k_events = 1, growth_rates = NULL, norm = FALSE,
    t_end=1000)

# Convert into TreeSE format.
tse <- TreeSummarizedExperiment(
                         assays = list(counts = x$counts), 
                         colData = x$time,
                         metadata = x$metadata)