Consumer-resource model simulation

Simulates a community time series using the consumer-resource model. The change rate of each species was defined as dx/dt = mumax*sum(monod)*X, where mumax is the vector of maximum growth rates for the species, monod is the monod growth rate, S/(Ks+S), where S is the concentration of the limiting resource, and Ks is the half-velocity constant for species X and S. X is the vector of abundances of species. The concentrations of resource will be set to 0 if they were calculated less than 0 during the iteration.

simulateConsumerResource(
  n_species,
  n_resources,
  eff = randomE(n_species, n_resources),
  consumers = runif(n = n_species, min = 0.1, max = 10),
  resources = runif(n = n_resources, min = 1, max = 100),
  mumax = rep(1, n_species),
  k_table = matrix(rgamma(n = n_species * n_resources, shape = 50, rate = 0.25), nrow =
    n_species),
  t_end = 1000,
  ...
)

Arguments

n_species

integer number of species

n_resources

interger number of resources

eff

a matrix of efficiency. How efficient are resources converted into biomass, negative values represent excreted resources (default: eff = randomE(n_species, n_resources))

consumers

numeric vector of species (default: consumers = runif(n = n_species, min = 0.1, max = 10))

resources

numeric vector of resources (default: resources = runif(n = n_resources, min = 1, max = 100))

mumax

numeric vector of maximum mu of species (default: mumax = rep(1, n_species))

k_table

a matrix of K values in monod model (default: k_table = matrix(rgamma(n=n_species*n_resources, shape = 50, rate = 0.25), nrow = n_species))

t_end

numeric scalar indicating the final time of the simulation (default: t_end = 1000)

...

additional parameters including 't_start', 't_step', and 't_store'

Value

an abundance matrix with species and resources abundance as rows and time points as columns

See also

Examples

x <- simulateConsumerResource(n_species = 2, n_resources = 4)