Generalized Lotka-Volterra (gLV) simulation

Simulates time series with the generalized Lotka-Volterra model.

simulateGLV(
  n_species,
  A,
  x = runif(n_species),
  b = runif(n_species),
  sigma_drift = 0.01,
  sigma_epoch = 0.3,
  sigma_external = 0.3,
  p_epoch = 0.01,
  t_external_events = c(12, 36, 48),
  t_external_durations = c(3, 10, 99),
  stochastic = FALSE,
  norm = FALSE,
  t_end = 1000,
  ...
)

Arguments

n_species

integer number of species

A

interaction matrix

x

numeric initial abundances

b

numeric growth rates

sigma_drift

numeric degree of drift (turnover of species) in each time step. (default: sigma_drift = 0.01)

sigma_epoch

numeric degree of epoch change of community (default: sigma_epoch = 0.3)

sigma_external

numeric degree of the external events/disturbances (default: sigma_external = 0.3)

p_epoch

numeric value of the probability/frequency of inherit periodic changes of community (default: p_epoch = 0.01)

t_external_events

numeric value of starting times of external events (default: t_external_events = c(12, 36, 48))

t_external_durations

numeric durations of external events (default: t_external_durations = c(3, 10, 99))

stochastic

logical scalar choosing whether the gLV model should be stochastic (default: stochastic = FALSE)

norm

logical scalar returning normalised abundances (proportions in each generation) (default: norm = FALSE)

t_end

numeric value of simulation end time (default: t_end = 1000)

...

additional parameters including 't_start', 't_step', and 't_store'

Value

simulateGLV returns an abundance matrix

Details

Simulates a community time series using the generalized Lotka-Volterra model, defined as dx/dt = x(b+Ax), where x is the vector of species abundances, diag(x) is a diagonal matrix with the diagonal values set to x. A is the interaction matrix and b is the vector of growth rates.

See also

Examples

A <- miaSim::powerlawA(4, alpha = 1.01)

x <- simulateGLV(n_species = 4, A, t_end = 1000)