Hubbell community simulation

Neutral species abundances simulation turned into SummarizedExperiment according to the Hubbell model.

simulateHubbell(N, M, I, d, m, tskip = 0, tend, norm = FALSE)

Arguments

N
Integer: the amount of different species initially in the local community

M
Integer: amount of different species in the metacommunity,including those of the local community

I
Integer: fixed amount of individuals in the local community (default: I = 1000)

d
Integer: fixed amount of deaths of local community individuals in each generation (default: d = 10)

m
Numeric: immigration rate: the probability that a death in the local community is replaced by a migrant of the metacommunity rather than by the birth of a local community member (default: m = 0.02)

tskip
Integer: number of generations that should not be included in the outputted species abundance matrix. (default: tskip = 0)

tend
Integer: number of simulations to be simulated

norm
Logical: whether the time series should be returned with the abundances as proportions (norm = TRUE) or the raw counts (norm = FALSE, default)

Value

simulateHubbell returns a SummarizedExperiment object

References

Rosindell, James et al. "The unified neutral theory of biodiversity and biogeography at age ten." Trends in ecology & evolution vol. 26,7 (2011).

Examples

col_data <- data.frame(sampleID = c(1:100),
                        stringsAsFactors = FALSE)

row_data <- data.frame(Kingdom = "A",
                        Phylum = rep(c("B1", "B2"), c(50, 50)),
                        Class = rep(c("C1", "C2"), each = 50),
                        ASV1 = paste0("D", 1:100),
                        ASV2 = paste0("E", 1:100),
                        ASV3 = paste0("F", 1:100),
                        ASV4 = paste0("G", 1:100),
                        ASV5 = paste0("H", 1:100),
                        ASV6 = paste0("J", 1:100),
                        ASV7 = paste0("K", 1:100),
                        row.names = rownames(paste0("species", seq_len(10))),
                        stringsAsFactors = FALSE)
row_data <- t(row_data)

simulateHubbell(N = 8, M = 10, I = 1000, d = 50, m = 0.02, tend = 100)

#> class: SummarizedExperiment 
#> dim: 10 100 
#> metadata(0):
#> assays(1): counts
#> rownames(10): 1 2 ... 9 10
#> rowData names(0):
#> colnames(100): t1 t2 ... t99 t100
#> colData names(0):