Hubbell's neutral model simulation applied to time series

Neutral species abundances simulation according to the Hubbell model. This model shows that losses in society can be replaced either by the birth of individuals or by immigration depending on their probabilities. The specific time between the events of birth or migration is calculated and time effect is considered to determine the next event.

simulateHubbellRates(
  community_initial,
  migration_p = 0.1,
  metacommunity_p = NULL,
  k_events = 1,
  growth_rates = NULL,
  norm = FALSE,
  t_end = 1000,
  list = TRUE,
  ...
)

Arguments

community_initial

numeric value a vector of integers, containing species counts greater or equal to zero.

migration_p

numeric immigration possibility. It defines the probability of migration when replacement is needed in the community. The value can be between 0 and 1. The sum of the probability of migration and the probability birth must be 1.

metacommunity_p

numeric value the probability of a species being found in the metacommunity.

k_events

integer number of steps performed at a time point. It can be equal or more than 1. Bigger k_events increases speed while decreasing precision.

growth_rates

numeric rate of change in community size.

norm

logical scalar choosing whether the time series should be returned with the abundances as proportions (norm = TRUE) or the raw counts (default: norm = FALSE)

t_end

numeric value of simulation end time (default: t_end = 1000)

list

logical scalar deciding whether output is a list object or not (default: norm = TRUE)

...

additional parameters including 't_start', 't_step', and 't_store'

Value

a community abundance matrix or a list object that contains growth rates, time points and metacommunity probabilities

References

Rosindell J, Hubbell SP, Etienne RS. The unified neutral theory of biodiversity and biogeography at age ten. Trends Ecol Evol. 2011 Jul;26(7):340-8. doi: 10.1016/j.tree.2011.03.024. Epub 2011 May 10. PMID: 21561679.

See also

Examples

x <- simulateHubbellRates(community_initial = c(0,5,10),
             migration_p = 0.01, metacommunity_p = NULL, k_events = 1,
             growth_rates = NULL, norm = FALSE, t_end=1000)