Generate time series with the Ricker model

The Ricker model is a discrete version of the generalized Lotka-Volterra model and is implemented here as proposed by Fisher and Mehta in PLoS ONE 2014.

simulateRicker(
  n_species,
  A,
  x = runif(n_species),
  K = runif(n_species),
  sigma = 0.05,
  explosion_bound = 10^8,
  tskip = 0,
  tend,
  norm = FALSE
)

Arguments

n_species

integer number of species

A

interaction matrix

x

numeric initial abundances

K

numeric carrying capacities

sigma

numeric value of noise level, if set to a non-positive value, no noise is added (default: sigma = 0.05)

explosion_bound

numeric value of boundary for explosion (default: explosion_bound = 10^8)

tskip

integer number of generations that should not be included in the outputted species abundance matrix (default: tskip = 0)

tend

integer number of simulations to be simulated

norm

logical scalar returning normalised abundances (proportions in each generation) (default: norm = FALSE)

Value

simulateRicker returns an abundance matrix with species abundance as rows and time points as columns

References

Fisher & Mehta (2014). Identifying Keystone Species in the Human Gut Microbiome from Metagenomic Timeseries using Sparse Linear Regression. PLoS One 9:e102451

See also

Examples

A <- powerlawA(10, alpha = 1.01)
x <- simulateRicker(n_species=10, A, tend=100)