Self-Organised Instability model (SOI) simulation

Generate time-series with The Self-Organised Instability (SOI) model. Implements a K-leap method for accelerating stochastic simulation.

simulateSOI(n_species, I, A, k = 5, com = NULL, tend, norm = FALSE)

Arguments

n_species

integer number of species

I

integer community size, number of available sites (individuals)

A

interaction matrix

k

integer number of transition events that are allowed to take place during one leap. (default: k = 5). Higher values reduce runtime, but also accuracy of the simulation.

com

a vector of initial community abundances If (default: com = NULL), based on migration rates

tend

integer timepoints to be returned in the time series (number of generations)

norm

logical scalar indicating whether the time series should be returned with the abundances as proportions (norm = TRUE) or the raw counts (default: norm = FALSE)

Value

abundance matrix consisting of species abundance as rows and time points as columns

See also

Examples

# Generate interaction matrix
A <- miaSim::powerlawA(10, alpha = 1.2)
# Simulate data from the SOI model
x <- simulateSOI(n_species = 10, I = 1000, A, k=5, com = NULL,
                                            tend = 150, norm = TRUE)