Generate simulation times and the indices of time points to return in simulation functions.

Generate simulation times and the indices of time points to return in simulation functions.

simulationTimes(t_start = 0, t_end = 1000, t_step = 0.1, t_store = 1000)

Arguments

t_start

Numeric scalar indicating the initial time of the simulation. (default: t_start = 0)

t_end

Numeric scalar indicating the final time of the dimulation (default: t_end = 1000)

t_step

Numeric scalar indicating the interval between simulation steps (default: t_step = 0.1)

t_store

Integer scalar indicating the number of evenly distributed time points to keep (default: t_store = 100)

Value

lists containing simulation times (t_sys) and the indices to keep.

Examples

Time <- simulationTimes(
    t_start = 0, t_end = 100, t_step = 0.5,
    t_store = 100
)
DefaultTime <- simulationTimes(t_end = 1000)